* add generic classes types  * update some deprecated methods  * Fix some potential null pointer exceptions  * annoting overloaded methods
 * update the minimum java version from 1.4 to 1.5
 * update the branch from 0.2.3 to 0.2.4




git-svn-id: file:///home/code/svn/ovd/trunk@943 2b2718be-6ee5-4ec5-ad1a-3fa14f9c7ca3

+++ applets/ovd-applet/src/com/sshtools/j2ssh/io/ByteArrayReader.java
@@ -111,7 +111,7 @@ public class ByteArrayReader
    */
   public static String readString(byte[] data, int start) {
     int len = (int) readInt(data, start);
-    byte[] chars = new byte[ (int) len];
+    byte[] chars = new byte[ len];
     System.arraycopy(data, start + 4, chars, 0, len);
 
     return new String(chars);
